+++ src/main/java/org/bytecamp19/seckill4/controller/MainController.java
@@ -36,7 +36,7 @@ public class MainController {
         this.orderService = orderService;
     }
 
-    private Session getSession(String sessionId, Integer uid) throws ForbiddenException {
+    private Session getSession(String sessionId, Long uid) throws ForbiddenException {
         Session session = sessionService.getSession(sessionId);
         logger.debug("session = " + session);
         if (session == null) {
@@ -49,7 +49,7 @@ public class MainController {
     }
 
     @GetMapping("product")
-    public Product getProduct(@Param("pid") Integer pid,
+    public Product getProduct(@Param("pid") Long pid,
                               @RequestHeader("sessionid") String sessionId) throws ForbiddenException {
         if (pid == null) {
             throw new ForbiddenException("pid not given");
@@ -68,8 +68,8 @@ public class MainController {
     public JSONObject placeOrder(@RequestBody JSONObject json,
                                  @RequestHeader("sessionid") String sessionId)
             throws ForbiddenException {
-        Integer pid = json.getInteger("pid");
-        Integer uid = json.getInteger("uid");
+        Long pid = json.getLong("pid");
+        Long uid = json.getLong("uid");
         if (uid == null || pid == null || sessionId == null || sessionId.isEmpty()) {
             throw new ForbiddenException("pid / uid / session not given");
         }
@@ -104,7 +104,7 @@ public class MainController {
     public JSONObject payOrder(@RequestBody JSONObject json,
                                @RequestHeader("sessionid") String sessionId)
             throws ForbiddenException {
-        Integer uid = json.getInteger("uid");
+        Long uid = json.getLong("uid");
         Integer price = json.getInteger("price");
         String orderId = json.getString("order_id");
         if (uid == null || price == null || orderId == null || sessionId == null || sessionId.isEmpty()) {
@@ -135,7 +135,7 @@ public class MainController {
     }
 
     @GetMapping("result")
-    public JSONObject getResult(@Param("uid") Integer uid,
+    public JSONObject getResult(@Param("uid") Long uid,
                                 @RequestHeader("sessionid") String sessionId)
             throws ForbiddenException {
         if (uid == null){
